% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_catalog}
\alias{get_catalog}
\title{Download GWAS catalog metadata}
\usage{
get_catalog()
}
\value{
Dataframe with current GWAS-Catalog summary statistics information (from the web). Takes around 2-3 min
}
\description{
Downloads cleaned metadata information from GWAS catalog website of all the studies including summary statistics
}
