% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ss.R
\name{get_ss}
\alias{get_ss}
\title{Download metadata of the current GWAS catalog information}
\usage{
get_ss(url.ss, out.dir, out.filename = NULL)
}
\arguments{
\item{url.ss}{The link of the summary statistics file from GWAS Catalog that you wish to download}

\item{out.dir}{Output directory where the file should be stored}

\item{out.filename}{Name to give for the downloaded file}
}
\value{
Download the file in the specified folder (out.dir) named as specified in out.filename
}
\description{
Downloads cleaned metadata information from GWAS catalog of all the studies only including summary statistics
}
\examples{
get_ss('http://ftp.ebi.ac.uk/pub/databases/gwas/summary_statistics/GCST90004001-GCST90005000/GCST90004618','output_dir','GCST90004618')
}
